Enabling the use on anytype field (including selectboxes)
Some CRM implementation use Selectbox (switching original vtiger inputbox) for some fields such as States / Countries, and on some additional vtiger extensiones can also require functionality on change on the address fields;

therefor the proposed changes are the following:
1. use input field name (not on element id)
2. Use jquery val() to change value
3. trigger .trigger('liszt:updated') & .trigger('change');

you have an awsome extensions and great that is is free! so i am giving my few improvments, it is working good on my installation so I hope you accept and push it to your next version!